refac: rename da camada de acesso a dados local Closes #70
