+++ ApiDAL/ApiDAL.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DBContextDAL\DbContextDAL.csproj" />
+    <ProjectReference Include="..\DBContextDAL\Repositories.csproj" />
     <ProjectReference Include="..\Models\Models.csproj" />
   </ItemGroup>
