+++ BLL/ApiOperationBaseBLL.cs
@@ -1,4 +1,5 @@
-﻿using DbContextDAL;
+﻿using Repositories;
+using Repositories.Interfaces;
 using Models.OperationQueue;
 
 namespace BLL
