+++ BLL/Books/Historic/Sync/BookHistoricSyncBLL.cs
@@ -1,5 +1,5 @@
 ﻿using BLL.Books.Historic.Interfaces;
-using DbContextDAL;
+using Repositories.Interfaces;
 using Models.Books.Historic;
 using Models.Exceptions;
