+++ BLL/User/UserService.cs
@@ -1,6 +1,6 @@
 ﻿using ApiDAL.Handlers;
 using ApiDAL.Interfaces;
-using DbContextDAL;
+using Repositories.Interfaces;
 using Models.Responses;
 using System.Text.Json.Nodes;
