+++ BLLTests/Books/Sync/BookSyncBLLTests.cs
@@ -1,5 +1,6 @@
 ﻿using ApiDAL.Interfaces;
-using DbContextDAL;
+using Repositories;
+using Repositories.Interfaces;
 using DBContextDAL;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Models.Books;
