--- Models/Books/Historic/BookHistoricList.cs
@@ -1,7 +1,0 @@
-﻿namespace Models.Books.Historic
-{
-    public class BookHistoricList
-    {
-        public required List<BookHistoric> List { get; set; }
-    }
-}
